[droidian-images] Allow specifying Droidian variants
Droidian 101 introduces support for "Droidian variants" in the
rdo backend.

To not be confused with the standard "variants" that are mostly
related to configuration/package sets/desktop environments that
are otherwise available in the standard repos (i.e. phosh/plasma),
Droidian variants are entirely different repositories which are
snapshotted _in parallel_ of the standard, or "base" repo.

Droidian variants allow even greater flexibility for device-specific
patches, and open other avenues for experimentation without affecting the
main repositories.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>